Create quizz.js
Through this script, we can use it to find answers to quizzes. To use this script, download the tampermonkey extension for google. Once the extension is loaded into google, click on it and then click the "Create a new script" and paste the above script onto the page. After your done, save it and the script is fully loaded and can be used in quizzez where the answers will be shown