Create quizz.py
u can inspect ur quizz's page and then just go to console and paste this code after that ur browser will give u a pop up in that u should wright the exact name from one of a member who is also in the quizz then the correct answer will be highlighted. this code is actually made by me and my friend, so ya the credit even goes to him.